+++ src/components/charts/LineChart.js
@@ -7,11 +7,7 @@ import EmptyPlaceHolder from '@Components/table/EmptyPlaceHolder'
 import { color } from '@src/assets/styles/variables'
 import TooltipBox from '@Components/tooltip/TooltipBox'
 import * as commonTag from '@Components/common/cdmCommon'
-
-const colorSet = {
-  blue: ['#d5e7fd', '#a5d2ff', '#63a3f3', '#3788ed', '#2f60c3', '#224b9f', '#1e3476', '#142352'],
-  green: ['#ceede7', '#97d9ce', '#24b7a3', '#0c8d84', '#006f75', '#00555a', '#043e4b', '#002340'],
-}
+import { getColorsByTheme } from '@Components/ChartColor'
 
 const LineChart = ({
   title,
@@ -20,9 +16,12 @@ const LineChart = ({
   yDataKey,
   theme,
   isPercent,
+  margin,
+  xData,
+  yData,
 }) => {
-  const colors = colorSet[theme] || colorSet.blue
   const newYDataKey = [].concat(yDataKey)
+  const colors = getColorsByTheme(theme, newYDataKey.length)
   const legendData = _.chain(newYDataKey)
     .map((entry, index) => ({ color: colors[index], text: entry }))
     .value()
@@ -35,23 +34,35 @@ const LineChart = ({
     return newValue
   }
 
-  const isEmpty = (items) => _.isEmpty(items)
-
   return (
     <div>
       <Heading size="18" style={{ marginBottom: '30px' }}>{title}</Heading>
-
       <commonTag.LegendList data={legendData} />
-
       {
-        isEmpty(data)
+        _.isEmpty(data)
           ? <EmptyPlaceHolder />
           : (
             <Rechart.ResponsiveContainer height={415}>
-              <Rechart.LineChart data={data} height={415}>
+              <Rechart.LineChart
+                data={data}
+                height={415}
+                margin={margin}
+              >
                 <Rechart.CartesianGrid vertical={false} stroke={color.$line_graph_xy_grey} />
-                <Rechart.XAxis tickLine={false} tickMargin={10} dataKey={xDataKey} stroke="rgba(0, 0, 0, 0.6)" />
-                <Rechart.YAxis axisLine={false} tickLine={false} tickFormatter={tickFormatter} tickMargin={10} stroke="rgba(0, 0, 0, 0.4)" />
+                <Rechart.XAxis tickLine={false} tickMargin={10} dataKey={xDataKey} stroke="rgba(0, 0, 0, 0.6)">
+                  {
+                    xData.label && (
+                      <Rechart.Label value={xData.label.value} offset={10} position="bottom" style={{ fill: 'rgba(0, 0, 0, 0.6)' }} />
+                    )
+                  }
+                </Rechart.XAxis>
+                <Rechart.YAxis axisLine={false} tickLine={false} tickFormatter={tickFormatter} tickMargin={10} stroke="rgba(0, 0, 0, 0.4)">
+                  {
+                    yData.label && (
+                      <Rechart.Label value={yData.label.value} offset={0} position="left" style={{ fill: 'rgba(0, 0, 0, 0.4)' }} />
+                    )
+                  }
+                </Rechart.YAxis>
                 <Rechart.Tooltip
                   isPercent={isPercent}
                   content={TooltipBox}
@@ -74,6 +85,11 @@ LineChart.defaultProps = {
   yDataKey: ['value', []],
   theme: 'blue',
   isPercent: false,
+  margin: {
+    top: 10, right: 5, bottom: 5, left: 5,
+  },
+  xData: {},
+  yData: {},
 }
 
 LineChart.propTypes = {
@@ -81,8 +97,28 @@ LineChart.propTypes = {
   data: PropTypes.arrayOf(PropTypes.shape({})),
   xDataKey: PropTypes.string,
   yDataKey: PropTypes.oneOfType([PropTypes.string, PropTypes.arrayOf(PropTypes.string)]),
-  theme: PropTypes.string,
+  theme: PropTypes.oneOf([
+    'blue', 'green', 'compare',
+    'theme-arrange-primary-sea', 'theme-arrange-secondary-teal', 'theme-arrange-tertiary-rose',
+    'theme-arrange-quaternary-gold', 'theme-arrange-quinary-berry',
+  ]),
   isPercent: PropTypes.bool,
+  margin: PropTypes.shape({
+    top: PropTypes.number,
+    right: PropTypes.number,
+    bottom: PropTypes.number,
+    left: PropTypes.number,
+  }),
+  xData: PropTypes.shape({
+    label: PropTypes.shape({
+      value: PropTypes.string.isRequired,
+    }),
+  }),
+  yData: PropTypes.shape({
+    label: PropTypes.shape({
+      value: PropTypes.string.isRequired,
+    }),
+  }),
 }
 
 export default LineChart