+++ src/index.js
@@ -42,6 +42,7 @@ import * as chartUtility from '@src/helper/chartUtility'
 import * as DateUtility from '@src/helper/DateUtility'
 
 import notifications from '@Components/notifications'
+import * as ChartColor from '@Components/ChartColor'
 
 export {
   BarGauge,
@@ -85,4 +86,5 @@ export {
   Pagination,
   DateUtility,
   notifications,
+  ChartColor,
 }